+++ AppleALC/kern_alc.cpp
@@ -279,6 +279,39 @@ IOService *AlcEnabler::gfxProbe(IOService *ctrl, IOService *provider, SInt32 *sc
 	return FunctionCast(gfxProbe, callbackAlc->orgGfxProbe)(ctrl, provider, score);
 }
 
+bool AlcEnabler::AppleHDAController_start(IOService* service, IOService* provider)
+{
+	uint32_t delay = 0;
+	if (PE_parse_boot_argn("alcdelay", &delay, sizeof(delay))) {
+		DBGLOG("alc", "found alc-delay override %u", delay);
+		provider->setProperty("alc-delay", &delay, sizeof(delay));
+	} else {
+		if (WIOKit::getOSDataValue(provider, "alc-delay", delay))
+			DBGLOG("alc", "found normal alc-delay %u", delay);
+	}
+	
+	if (delay > 3000) {
+		SYSLOG("alc", "alc delay cannot exceed 3000 ms, ignore it");
+		delay = 0;
+	}
+		
+	if (delay != 0) {
+		DBGLOG("alc", "delay AppleHDAController::start for %d ms", delay);
+		IOSleep(delay);
+	}
+	return FunctionCast(AppleHDAController_start, callbackAlc->orgAppleHDAController_start)(service, provider);
+}
+
+#ifdef DEBUG
+IOReturn AlcEnabler::IOHDACodecDevice_executeVerb(void *that, uint16_t a1, uint16_t a2, uint16_t a3, unsigned int *a4, bool a5)
+{
+	IOReturn result = FunctionCast(IOHDACodecDevice_executeVerb, callbackAlc->orgIOHDACodecDevice_executeVerb)(that, a1, a2, a3, a4, a5);
+	if (result != KERN_SUCCESS)
+		DBGLOG("alc", "IOHDACodecDevice::executeVerb with parameters a1 = %u, a2 = %u, a3 = %u failed with result = %x", a1, a2, a3, result);
+	return result;
+}
+#endif
+
 uint32_t AlcEnabler::getAudioLayout(IOService *hdaDriver) {
 	auto parent = hdaDriver->getParentEntry(gIOServicePlane);
 	uint32_t layout = 0;
@@ -573,6 +606,20 @@ void AlcEnabler::processKext(KernelPatcher &patcher, size_t index, mach_vm_addre
 		// patch AppleHDA to remove redundant logs
 		if (!ADDPR(debugEnabled))
 			eraseRedundantLogs(patcher, kextIndex);
+	}
+	
+#ifdef DEBUG
+	if (ADDPR(debugEnabled) && !(progressState & ProcessingState::PatchHDAFamily) && kextIndex == KextIdIOHDAFamily) {
+		progressState |= ProcessingState::PatchHDAFamily;
+		KernelPatcher::RouteRequest request("__ZN16IOHDACodecDevice11executeVerbEtttPjb", IOHDACodecDevice_executeVerb, orgIOHDACodecDevice_executeVerb);
+		patcher.routeMultiple(index, &request, 1, address, size);
+	}
+#endif
+	
+	if (!(progressState & ProcessingState::PatchHDAController) && kextIndex == KextIdAppleHDAController) {
+		progressState |= ProcessingState::PatchHDAController;
+		KernelPatcher::RouteRequest request("__ZN18AppleHDAController5startEP9IOService", AppleHDAController_start, orgAppleHDAController_start);
+		patcher.routeMultiple(index, &request, 1, address, size);
 	}
 	
 	// Ignore all the errors for other processors